Refactor Stripe tests mocks
