+++ apps/stripe/src/__tests__/mocks/saleor-events/transaction-cancelation-request-event.ts
@@ -1,7 +1,7 @@
 import { TransactionCancelationRequestedEventFragment } from "@/generated/graphql";
 
-import { mockedSaleorChannelId } from "./constants";
-import { mockedStripePaymentIntentId } from "./mocked-stripe-payment-intent-id";
+import { mockedSaleorChannelId } from "../constants";
+import { mockedStripePaymentIntentId } from "../mocked-stripe-payment-intent-id";
 
 export const getMockedTransactionCancelationRequestedEvent =
   (): TransactionCancelationRequestedEventFragment => ({
