+++ apps/stripe/src/app/api/webhooks/saleor/transaction-cancelation-requested/use-case.test.ts
@@ -7,7 +7,7 @@ import { mockedSaleorAppId } from "@/__tests__/mocks/constants";
 import { mockedStripePaymentIntentId } from "@/__tests__/mocks/mocked-stripe-payment-intent-id";
 import { mockedStripePaymentIntentsApi } from "@/__tests__/mocks/mocked-stripe-payment-intents-api";
 import { mockedSaleorApiUrl } from "@/__tests__/mocks/saleor-api-url";
-import { getMockedTransactionCancelationRequestedEvent } from "@/__tests__/mocks/transaction-cancelation-request-event";
+import { getMockedTransactionCancelationRequestedEvent } from "@/__tests__/mocks/saleor-events/transaction-cancelation-request-event";
 import {
   AppIsNotConfiguredResponse,
   BrokenAppResponse,
