+++ apps/stripe/src/app/api/webhooks/stripe/use-case-success.test.ts
@@ -3,7 +3,7 @@ import { ok } from "neverthrow";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
 import { mockedAppConfigRepo } from "@/__tests__/mocks/app-config-repo";
-import { mockAdyenWebhookUrl, mockedSaleorTransactionIdBranded } from "@/__tests__/mocks/constants";
+import { mockAdyenWebhookUrl, mockedSaleorTransactionId } from "@/__tests__/mocks/constants";
 import { mockAuthData } from "@/__tests__/mocks/mock-auth-data";
 import { mockedStripePaymentIntentId } from "@/__tests__/mocks/mocked-stripe-payment-intent-id";
 import { MockedTransactionRecorder } from "@/__tests__/mocks/mocked-transaction-recorder";
@@ -74,7 +74,7 @@ describe("StripeWebhookUseCase - handling payment_intent.success event", () => {
 
     mockTransactionRecorder.transactions = {
       [stripePiId]: new RecordedTransaction({
-        saleorTransactionId: mockedSaleorTransactionIdBranded,
+        saleorTransactionId: mockedSaleorTransactionId,
         stripePaymentIntentId: stripePiId,
         saleorTransactionFlow: createSaleorTransactionFlow("CHARGE"),
         resolvedTransactionFlow: createResolvedTransactionFlow("CHARGE"),
@@ -141,7 +141,7 @@ describe("StripeWebhookUseCase - handling payment_intent.success event", () => {
 
     mockTransactionRecorder.transactions = {
       [stripePiId]: new RecordedTransaction({
-        saleorTransactionId: mockedSaleorTransactionIdBranded,
+        saleorTransactionId: mockedSaleorTransactionId,
         stripePaymentIntentId: stripePiId,
         saleorTransactionFlow: createSaleorTransactionFlow("AUTHORIZATION"),
         resolvedTransactionFlow: createResolvedTransactionFlow("AUTHORIZATION"),
@@ -227,7 +227,7 @@ describe("StripeWebhookUseCase - handling payment_intent.amount_capturable_updat
 
     mockTransactionRecorder.transactions = {
       [stripePiId]: new RecordedTransaction({
-        saleorTransactionId: mockedSaleorTransactionIdBranded,
+        saleorTransactionId: mockedSaleorTransactionId,
         stripePaymentIntentId: stripePiId,
         saleorTransactionFlow: createSaleorTransactionFlow("AUTHORIZATION"),
         resolvedTransactionFlow: createResolvedTransactionFlow("AUTHORIZATION"),
@@ -313,7 +313,7 @@ describe("StripeWebhookUseCase - handling payment_intent.payment_failed event", 
 
     mockTransactionRecorder.transactions = {
       [stripePiId]: new RecordedTransaction({
-        saleorTransactionId: mockedSaleorTransactionIdBranded,
+        saleorTransactionId: mockedSaleorTransactionId,
         stripePaymentIntentId: stripePiId,
         saleorTransactionFlow: createSaleorTransactionFlow("CHARGE"),
         resolvedTransactionFlow: createResolvedTransactionFlow("CHARGE"),
@@ -380,7 +380,7 @@ describe("StripeWebhookUseCase - handling payment_intent.payment_failed event", 
 
     mockTransactionRecorder.transactions = {
       [stripePiId]: new RecordedTransaction({
-        saleorTransactionId: mockedSaleorTransactionIdBranded,
+        saleorTransactionId: mockedSaleorTransactionId,
         stripePaymentIntentId: stripePiId,
         saleorTransactionFlow: createSaleorTransactionFlow("AUTHORIZATION"),
         resolvedTransactionFlow: createResolvedTransactionFlow("AUTHORIZATION"),
@@ -465,7 +465,7 @@ describe("StripeWebhookUseCase - handling payment_intent.processing event", () =
 
     mockTransactionRecorder.transactions = {
       [stripePiId]: new RecordedTransaction({
-        saleorTransactionId: mockedSaleorTransactionIdBranded,
+        saleorTransactionId: mockedSaleorTransactionId,
         stripePaymentIntentId: stripePiId,
         saleorTransactionFlow: createSaleorTransactionFlow("CHARGE"),
         resolvedTransactionFlow: createResolvedTransactionFlow("CHARGE"),
@@ -530,7 +530,7 @@ describe("StripeWebhookUseCase - handling payment_intent.processing event", () =
 
     mockTransactionRecorder.transactions = {
       [stripePiId]: new RecordedTransaction({
-        saleorTransactionId: mockedSaleorTransactionIdBranded,
+        saleorTransactionId: mockedSaleorTransactionId,
         stripePaymentIntentId: stripePiId,
         saleorTransactionFlow: createSaleorTransactionFlow("AUTHORIZATION"),
         resolvedTransactionFlow: createResolvedTransactionFlow("AUTHORIZATION"),
@@ -613,7 +613,7 @@ describe("StripeWebhookUseCase - handling payment_intent.requires_action event",
 
     mockTransactionRecorder.transactions = {
       [stripePiId]: new RecordedTransaction({
-        saleorTransactionId: mockedSaleorTransactionIdBranded,
+        saleorTransactionId: mockedSaleorTransactionId,
         stripePaymentIntentId: stripePiId,
         saleorTransactionFlow: createSaleorTransactionFlow("CHARGE"),
         resolvedTransactionFlow: createResolvedTransactionFlow("CHARGE"),
@@ -680,7 +680,7 @@ describe("StripeWebhookUseCase - handling payment_intent.requires_action event",
 
     mockTransactionRecorder.transactions = {
       [stripePiId]: new RecordedTransaction({
-        saleorTransactionId: mockedSaleorTransactionIdBranded,
+        saleorTransactionId: mockedSaleorTransactionId,
         stripePaymentIntentId: stripePiId,
         saleorTransactionFlow: createSaleorTransactionFlow("AUTHORIZATION"),
         resolvedTransactionFlow: createResolvedTransactionFlow("AUTHORIZATION"),
@@ -765,7 +765,7 @@ describe("StripeWebhookUseCase - handling payment_intent.canceled event", () => 
 
     mockTransactionRecorder.transactions = {
       [stripePiId]: new RecordedTransaction({
-        saleorTransactionId: mockedSaleorTransactionIdBranded,
+        saleorTransactionId: mockedSaleorTransactionId,
         stripePaymentIntentId: stripePiId,
         saleorTransactionFlow: createSaleorTransactionFlow("CHARGE"),
         resolvedTransactionFlow: createResolvedTransactionFlow("CHARGE"),
@@ -830,7 +830,7 @@ describe("StripeWebhookUseCase - handling payment_intent.canceled event", () => 
 
     mockTransactionRecorder.transactions = {
       [stripePiId]: new RecordedTransaction({
-        saleorTransactionId: mockedSaleorTransactionIdBranded,
+        saleorTransactionId: mockedSaleorTransactionId,
         stripePaymentIntentId: stripePiId,
         saleorTransactionFlow: createSaleorTransactionFlow("AUTHORIZATION"),
         resolvedTransactionFlow: createResolvedTransactionFlow("AUTHORIZATION"),
@@ -912,7 +912,7 @@ describe("StripeWebhookUseCase - handling charge.refund.updated event", () => {
 
     mockTransactionRecorder.transactions = {
       [mockedStripePaymentIntentId]: new RecordedTransaction({
-        saleorTransactionId: mockedSaleorTransactionIdBranded,
+        saleorTransactionId: mockedSaleorTransactionId,
         stripePaymentIntentId: mockedStripePaymentIntentId,
         saleorTransactionFlow: createSaleorTransactionFlow("CHARGE"),
         resolvedTransactionFlow: createResolvedTransactionFlow("CHARGE"),
@@ -980,7 +980,7 @@ describe("StripeWebhookUseCase - handling charge.refund.updated event", () => {
 
     mockTransactionRecorder.transactions = {
       [mockedStripePaymentIntentId]: new RecordedTransaction({
-        saleorTransactionId: mockedSaleorTransactionIdBranded,
+        saleorTransactionId: mockedSaleorTransactionId,
         stripePaymentIntentId: mockedStripePaymentIntentId,
         saleorTransactionFlow: createSaleorTransactionFlow("CHARGE"),
         resolvedTransactionFlow: createResolvedTransactionFlow("CHARGE"),
@@ -1048,7 +1048,7 @@ describe("StripeWebhookUseCase - handling charge.refund.updated event", () => {
 
     mockTransactionRecorder.transactions = {
       [mockedStripePaymentIntentId]: new RecordedTransaction({
-        saleorTransactionId: mockedSaleorTransactionIdBranded,
+        saleorTransactionId: mockedSaleorTransactionId,
         stripePaymentIntentId: mockedStripePaymentIntentId,
         saleorTransactionFlow: createSaleorTransactionFlow("CHARGE"),
         resolvedTransactionFlow: createResolvedTransactionFlow("CHARGE"),
