+++ apps/stripe/src/modules/transactions-recording/repositories/dynamodb/dynamodb-transaction-recorder-repo.test.ts
@@ -2,7 +2,7 @@ import { DynamoDBDocumentClient, GetCommand, PutCommand } from "@aws-sdk/lib-dyn
 import { mockClient } from "aws-sdk-client-mock";
 import { beforeEach, describe, expect, it } from "vitest";
 
-import { mockedSaleorAppId, mockedSaleorTransactionIdBranded } from "@/__tests__/mocks/constants";
+import { mockedSaleorAppId, mockedSaleorTransactionId } from "@/__tests__/mocks/constants";
 import { getMockedRecordedTransaction } from "@/__tests__/mocks/mocked-recorded-transaction";
 import { mockedStripePaymentIntentId } from "@/__tests__/mocks/mocked-stripe-payment-intent-id";
 import { mockedSaleorApiUrl } from "@/__tests__/mocks/saleor-api-url";
@@ -89,7 +89,7 @@ describe("DynamoDBTransactionRecorderRepo", () => {
             PK: "https://foo.bar.saleor.cloud/graphql/#saleor-app-id",
             SK: "TRANSACTION#pi_TEST_TEST_TEST",
             paymentIntentId: mockedStripePaymentIntentId,
-            saleorTransactionId: mockedSaleorTransactionIdBranded,
+            saleorTransactionId: mockedSaleorTransactionId,
             saleorTransactionFlow: "CHARGE",
             resolvedTransactionFlow: "CHARGE",
             selectedPaymentMethod: "card",
