+++ apps/stripe/src/modules/transactions-recording/repositories/dynamodb/recorded-transaction-db-model.test.ts
@@ -1,7 +1,7 @@
 import { Parser } from "dynamodb-toolbox";
 import { describe, expect, it } from "vitest";
 
-import { mockedSaleorAppId, mockedSaleorTransactionIdBranded } from "@/__tests__/mocks/constants";
+import { mockedSaleorAppId, mockedSaleorTransactionId } from "@/__tests__/mocks/constants";
 import { mockAuthData } from "@/__tests__/mocks/mock-auth-data";
 import { mockedStripePaymentIntentId } from "@/__tests__/mocks/mocked-stripe-payment-intent-id";
 import { mockedSaleorApiUrl } from "@/__tests__/mocks/saleor-api-url";
@@ -30,8 +30,8 @@ describe("DynamoDbRecordedTransaction", () => {
       expect(() =>
         parser.parse({
           PK: `${mockedSaleorApiUrl}#${mockedSaleorAppId}`,
-          SK: `TRANSACTION#${mockedSaleorTransactionIdBranded}`,
-          saleorTransactionId: mockedSaleorTransactionIdBranded,
+          SK: `TRANSACTION#${mockedSaleorTransactionId}`,
+          saleorTransactionId: mockedSaleorTransactionId,
           saleorTransactionFlow: "AUTHORIZATION",
           resolvedTransactionFlow: "CHARGE",
           selectedPaymentMethod: "card",
@@ -41,9 +41,9 @@ describe("DynamoDbRecordedTransaction", () => {
       expect(
         parser.parse({
           PK: `${mockedSaleorApiUrl}#${mockedSaleorAppId}`,
-          SK: `TRANSACTION#${mockedSaleorTransactionIdBranded}`,
+          SK: `TRANSACTION#${mockedSaleorTransactionId}`,
           paymentIntentId: mockedStripePaymentIntentId,
-          saleorTransactionId: mockedSaleorTransactionIdBranded,
+          saleorTransactionId: mockedSaleorTransactionId,
           saleorTransactionFlow: "AUTHORIZATION",
           resolvedTransactionFlow: "CHARGE",
           selectedPaymentMethod: "card",
